+++ sentry-android-core/src/main/java/io/sentry/android/core/ActivityLifecycleIntegration.java
@@ -388,6 +388,12 @@ public final class ActivityLifecycleIntegration
   }
 
   @TestOnly
+  @NotNull
+  ActivityFramesTracker getActivityFramesTracker() {
+    return activityFramesTracker;
+  }
+
+  @TestOnly
   @Nullable
   ISpan getAppStartSpan() {
     return appStartSpan;